+++ lib/glimmer/swt/command_handlers/property_command_handler.rb
@@ -8,10 +8,10 @@ module Glimmer
 
         def can_handle?(parent, command_symbol, *args, &block)
           parent.respond_to?(:set_attribute) and
-          parent.respond_to?(:has_attribute?) and
-          args.size > 0 and
-          block == nil and
-          parent.has_attribute?(command_symbol, *args)
+            parent.respond_to?(:has_attribute?) and
+            args.size > 0 and
+            block == nil and
+            parent.has_attribute?(command_symbol, *args)
         end
 
         def do_handle(parent, command_symbol, *args, &block)